push: allow setting commit date
If you want to preserve the commit hash across multiple guilt pushes of the
same patch set, the patch set needs to have a fixed commit date. This patch
allows guilt to pass the --committer-date-is-author-date flag to git, thus
setting the commit date to author date, which can be stored in the "From:"
header in the patch file.

Signed-off-by: Viacheslav Galaktionov <slavka@slavkin.site>